Vote result now only accessible by admins
