+++ src/main/java/com/arcbees/bourseje/client/api/VoteService.java
@@ -16,13 +16,10 @@
 
 package com.arcbees.bourseje.client.api;
 
-import java.util.Collection;
-
 import javax.ws.rs.GET;
 import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 
-import com.arcbees.bourseje.shared.CandidateResult;
 import com.arcbees.bourseje.shared.ResourcesPath;
 import com.arcbees.bourseje.shared.VoteItem;
 import com.arcbees.bourseje.shared.VoteState;
@@ -33,10 +30,6 @@ public interface VoteService {
     @GET
     @Path(ResourcesPath.CURRENT_VOTE_STATE)
     RestAction<VoteState> getCurrentVoteState();
-
-    @GET
-    @Path(ResourcesPath.NUMBER_OF_VOTES)
-    RestAction<Collection<CandidateResult>> getVotesPerCandidate();
 
     @POST
     @Path(ResourcesPath.CODE)
